# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Task Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Book": {
        "type": "object",
        "required": [
          "title",
          "author",
          "available",
          "updatedEdition"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the book",
            "example": "The Book"
          },
          "author": {
            "type": "string",
            "description": "Name of the book's author",
            "example": "David"
          },
          "available": {
            "type": "boolean",
            "description": "Whether the book is currently available",
            "example": true
          },
          "updatedEdition": {
            "type": "boolean",
            "description": "Indicates if the book is the latest or updated edition",
            "example": false
          }
        }
      },
...
      "Review": {
        "type": "object",
        "required": [
          "id",
          "bookId",
          "userId",
          "rating",
          "comment"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "review123"
          },
          "bookId": {
            "type": "string",
            "example": "book123"
          },
          "userId": {
            "type": "string",
            "example": "user456"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "comment": {
            "type": "string",
            "example": "One of the best books!"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "CreateReviewRequest": {
        "type": "object",
        "required": [
          "bookId",
          "userId",
          "rating",
          "comment"
        ],
        "properties": {
          "bookId": {
            "type": "string",
            "example": "book123"
          },
          "userId": {
            "type": "string",
            "example": "user456"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "comment": {
            "type": "string",
            "example": "One of the best books I've read this year!"
          }
        },
        "description": "Create a review"
      },
      "UpdateReviewRequest": {
        "type": "object",
        "description": "Fields of a review",
        "properties": {
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 3
          },
          "comment": {
            "type": "string",
            "example": "Updated review: after rereading, some parts felt slow."
          }
        },
        "minProperties": 1
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/books": {
      "post": {
        "summary": "Create a new book",
        "tags": [
          "Books"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "author",
                  "available"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Clean Code"
                  },
                  "author": {
                    "type": "string",
                    "example": "Robert C. Martin"
                  },
                  "available": {
                    "type": "boolean",
                    "example": true
                  },
                  "updatedEdition": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book created"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      },
      "get": {
        "summary": "Retrieve all books",
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Books retrieved"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Book"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "summary": "Retrieve a book by ID",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the book"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the book",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book found"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "summary": "Update a book by ID",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the book to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "author",
                  "available"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Clean Code (Updated Edition)"
                  },
                  "author": {
                    "type": "string",
                    "example": "Robert C. Martin"
                  },
                  "available": {
                    "type": "boolean",
                    "example": true
                  },
                  "updatedEdition": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book updated"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a book by ID",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the book to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book deleted"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    },
...
    "/api/v1/reviews": {
      "post": {
        "summary": "Create a new review for a book",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Review created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Only User, Librarian, or Admin can create reviews"
          }
        }
      }
    },
    "/api/v1/books/{bookId}/reviews": {
      "get": {
        "summary": "Retrieve all reviews for a specific book",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the book to get reviews for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved book reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Reviews retrieved"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Review"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only Admin or Librarian can view all reviews"
          }
        }
      }
    },
    "/api/v1/reviews/{id}": {
      "put": {
        "summary": "Update a review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the review to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Only User, Librarian, or Admin may update reviews"
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the review to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully"
          },
          "403": {
            "description": "User, Admin or Librarian may delete reviews"
          },
          "404": {
            "description": "Review not found"
          }
        }
      }
    },
...
  },
...
}